requirements.txt
beautifulsoup4
pandas
opencsv
openpyxl
pytest-playwright
pyarrow
